Random Password Generator
This mini project is a Python-based tool that generates a secure, random password of n length for use . The password consists of a mix of uppercase letters, lowercase letters, numbers, and special characters to ensure strong security. The user can specify the desired password length, and the script will generate a unique and secure password accordingly. This project utilizes the random and string modules in Python and can be extended with a CLI or GUI for enhanced usability.